Статистика вынесена в отдельный компонент
