+++ src/types/treeStatistics.ts
@@ -1,0 +1,5 @@
+import { ITree } from "./tree";
+
+export interface TreeStatisticsProps {
+  tree: ITree;
+}
